Replace errant phoenix references with superteams
This app was originally and mistakenly named `phoenix`, which resulted
in a huge number of bugs.

The fastest and safest way to fix this is to destroy the repo and start
again. I'm still not certain that I found and fixed all the bugs caused
by this issue.

This seems to at least run the app now. Who knows what else is broken,
though.