Re-add bootstrap
I admire the desire to start fresh without bootstrap, but right now the
app is basically unusable without these starting styles. When you're
ready to get rid of it, get rid of it and style your app on your own in
the same commit, or at least the same branch. Do it as one thought.